djangoProject3/requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1

djangorestframework==3.11.0
djoser==2.0.3
django-cors-headers==3.1.1
